==> notebooks/1.TransitScrapper.ipynb <==
#### Extracción tabla datos Selenium

import pandas as pd
from selenium.webdriver.chrome.options import Options
from webdriver_manager.chrome import ChromeDriverManager

#driver configuration
opciones=Options()

opciones.add_experimental_option('excludeSwitches', ['enable-automation'])
opciones.add_experimental_option('useAutomationExtension', False)
opciones.headless=False    # si True, no aperece la ventana (headless=no visible)
opciones.add_argument('--start-maximized')         # comienza maximizado
#opciones.add_argument('user-data-dir=selenium')    # mantiene las cookies
#opciones.add_extension('driver_folder/adblock.crx')       # adblocker
opciones.add_argument('--incognito')
from selenium.webdriver.common.by import By
from IPython.display import Image # Para mostrar imágenes en python
from bs4 import BeautifulSoup as bs
import time
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.common.action_chains import ActionChains

pd.set_option('display.max_columns', None)

from selenium import webdriver
PATH = 'driver/chromedriver.exe'
driver = webdriver.Chrome()

url = 'https://www.portsdebalears.com/es/buques-en-puerto' # Definimos url de búsqueda

driver.get(url)

driver.find_element(By.XPATH, '//*[@id="popup-buttons"]/button[1]').click() # Clickamos en cookies

# Ojo, la tabla que queremos extraer está dentro de un frame distinto al de la página principal (por eso la url no cambia al buscar tablas de datos diferentes)

driver.switch_to.frame(driver.find_element(By.XPATH, '//*[@id="giswebframe"]'))

driver.find_element(By.ID, 'trafico_menu').click() # Ahora si podemos clickar sobre los datos históricos

# Me defino una función que me haga el triple click

actions = ActionChains(driver)
def triple_click(element_x):
    actions.click(element_x).click(element_x).click(element_x).perform()

/html/body/div[4]/div[2]/div[3]/div/div[1]/div[2]/input
/html/body/div[4]/div[2]/div[3]/div/div[1]/div[4]/input

# Nuestro próximo reto es introducir las fechas que queremos de rango
# Tendremos que clickar en la caja y, mejor que clickar sobre fecha, introducir por texto el rango de fechas que nos interesa

fecha_ini = driver.find_element(By.XPATH, '/html/body/div[4]/div[2]/div[3]/div/div[1]/div[2]/input') # Apunto con el full XPATH
triple_click(fecha_ini)
fecha_ini.send_keys("01/06/2023") # Meto fecha entrada
time.sleep(2)

fecha_fin = driver.find_element(By.XPATH, '/html/body/div[4]/div[2]/div[3]/div/div[1]/div[4]/input') # Apunto con el full XPATH
triple_click(fecha_fin)
fecha_fin.send_keys("31/08/2023") # Meto fecha salida
time.sleep(2)
driver.find_element(By.XPATH, '/html/body/div[4]/div[2]/div[3]/div/div[1]/div[5]/button[1]/span').click()
time.sleep(10)

/html/body/div[4]/div[8]/div/div[2]/div/div/div/div[2]/div[3]/div[3]/div/table/tbody

Habiendo introducido las fechas vamos a obtener la tabla de datos de transito de buques

time.sleep(5)
tabla = driver.find_element(By.XPATH, '/html/body/div[4]/div[8]/div/div[2]/div/div/div/div[2]/div[3]/div[3]/div/table/tbody') # Asignamos variable tabla al cuerpo de la tabla
filas = tabla.find_elements(By.TAG_NAME, 'tr') # Cada fila está metida en un 'tr'

data = []

for f in filas:

    elementos = f.find_elements(By.TAG_NAME, 'td') # Cada fila está metida en un 'td'

    tmp = []

    for e in elementos:
        tmp.append(e.text)

    data.append(tmp)

df = pd.DataFrame(data)

df.head()

df_transito = df.copy()

df_transito.head()

df_transito.shape

df_transito.info()

df_transito[9].unique() # Nombres de barcos únicos

len(df_transito[9].unique())

90 barcos únicos de entre 1050 incidencias de tránsito.

df_transito[18].unique() # Empresas dueñas de los barcos 

len(df_transito[18].unique())

De las 1050 incidencias de barcos, todos se pueden catalogar en 11 empresas.

#### Limpieza dataframe

# Podemos borrar con toda tranquilidad las columnas [0:8] y [24::] porque no contienen información.

df_transito = df_transito.drop(df_transito.columns[0:9], axis=1) 

df_transito

df_transito.reset_index(drop=True)

df_transito.columns = [str(i) for i in range(24)]

df_transito.columns

df_transito.head()

# Voy a crear una lista con las columnas que quiero borrar y hago drop de ellas
df_transito.drop(columns = ['23'], inplace=True)
df_transito.head()



df_transito.info()

# Voy a renombrar las columnas a sus respectivos heads
heads = ['barco', 'origen', 'destino', 'llegada', 'salida', 'alineacion', 'consignatorio', 'gt','escala', 'bandera', 'eslora', 'calado']

df_transito.columns = heads

df_transito.head()

df_transito.drop_duplicates()

df_clean = df_transito.copy()
df_clean.head()

Perfecto. Tenemos un dataframe limpio con todas las incidencias de buques que han llegado y salido de puertos en el Mediterráneo entre los meses de junio y agosto junto con su información.
* Procedemos a pasar el dataframe a un csv para almacenarlo.

df_clean.to_csv('../data/transit_clean.csv', index=False)

# Voy a bajarme el csv de nuevo y pasarlo con indices de cara a que en SQL cada tabla tenga su indice y así poder usarlos de PK's.
df_index = pd.read_csv('../data/transit_clean.csv')
df_index.head()

df_index = df_index.drop(df_index.index[0])
df_index.head()

# Tambien voy a ponerle nombre a la columna del id -> transit_id
df_index = df_index.rename(columns={'Unnamed: 0':'transit_id'})
df_index.head()

df_index.to_csv('../data/transit_clean.csv', index=False)

driver.quit()
