==> notebooks/3.OTLE_csv.ipynb <==
#### OTLE 

Este notebook sirve para tener un preview del fichero CSV obtenido del Observatorio de Transporte y Logística de España (OTLE).

Se ha descargado el archivo csv en el siguiente [link](https://apps.fomento.gob.es/BDOTLE/visorBDpop.aspx?i=328), bajo el siguiente título: *Tráfico de mercancía y pasaje en puertos de interés general (número de buques y GT) por Autoridad Portuaria, tipo de buque y nacionalidad*


import pandas as pd
pd.set_option('display.max_columns', None)

df = pd.read_csv('../data/Num_buques_GT_Tipo_Buque.csv')
df.head()

df.info()

Tomando algo de detalle a las columnas desde la 'gt' en adelante y al [link](https://apps.fomento.gob.es/BDOTLE/visorBDpop.aspx?i=328) de la fuente se puede proceder a borrar todas estas columnas. Pasar a renombrarlo es un proceso lento y que cobra poco sentido cuando con simples queries se pueden hallar los sumatorios de campos previos.

df = df.iloc[:, :6]
df.head()

df = df.dropna()
df.info()

df.to_csv('../data/otle_gt.csv', index=True)

Pasamos el nuevo y limpio dataframe a csv con título **otle_gt.csv**. Después de este paso podemos crear el entorno de la base de datos en SQL y posteriormente cargar las tablas con el contenido de dichos csv's.
