==> notebooks/4.sql_builder.ipynb <==
#### Creación base de datos - SQL
Este notebook se va a utilizar para crear y rellenar la base de datos con las tablas previamente trabajadas.

import mysql.connector as conn
from sqlalchemy import create_engine
import warnings
warnings.filterwarnings('ignore')
import pandas as pd

with open('contraseñas.txt') as file:
    mypasswd = file.read()

# Nos conectamos al servidor 
conexion = conn.connect(host='localhost',
                        user='root',
                        passwd=f'{mypasswd}'
                        )

cursor  = conexion.cursor()

cursor

cursor.execute('drop database if exists transito')

cursor.execute('create database transito;')

conexion = conn.connect(host='localhost',
                        user='root',
                        passwd=f'{mypasswd}',
                        database='transito'
                        )

cursor  = conexion.cursor()

c = cursor.execute
c

#### Insertamos tablas en el database

# Leemos los 4 archivos csv que nos interesan
df_transit_clean = pd.read_csv('../data/transit_clean.csv')
df_puertos = pd.read_csv('../data/puertos.csv')
df_transit_plus_weather = pd.read_csv('../data/transit_plus_weather.csv')
df_otle_gt = pd.read_csv('../data/otle_gt.csv')

#### Transit clean

df_transit_clean.head()

c('drop table if exists transit_clean;')

query ='''

create table transit_clean(
    transit_id int primary key not null,
    barco varchar(45),
    origen varchar(45),
    destino	varchar(45),
    llegada	varchar(45),
    salida varchar(45),
    alineacion varchar(100),
    consignatorio varchar(100),
    gt float,
    escala float,
    bandera varchar(45),
    eslora float,
    calado float
    );
'''

c(query)

# Relleno datos
nombre_tabla = 'transit_clean'
nombre_columnas = ','.join(df_transit_clean.columns) # creo una string

#bucle por filas
for i in range(df_transit_clean.shape[0]):
    valores = tuple(df_transit_clean.iloc[i].values)

    insert_query = f'insert into {nombre_tabla} ({nombre_columnas}) values {valores}'

    cursor.execute(insert_query)


insert_query

# Comprobamos

cursor.execute('select * from transit_clean limit 2;')

for x in cursor:
    print(x)

# Comprobamos tamaño
# check 

cursor.execute('select count(*) from transit_clean;')

for x in cursor:
    print(x)

#### Puertos

df_puertos = df_puertos.rename(columns={'transit_coords_id': 'puertos_id'})
df_puertos.head()

c('drop table if exists puertos;')

query ='''

create table puertos(
    puertos_id int primary key not null,
    Puerto varchar(100),
    Latitud float,
    Longitud float
    );
'''

c(query)

# Relleno datos
nombre_tabla = 'puertos'
nombre_columnas = ','.join(df_puertos.columns) # creo una string

#bucle por filas
for i in range(df_puertos.shape[0]):
    valores = tuple(df_puertos.iloc[i].values)

    insert_query = f'insert into {nombre_tabla} ({nombre_columnas}) values {valores}'

    cursor.execute(insert_query)

insert_query

# Comprobamos

cursor.execute('select * from puertos limit 2;')

for x in cursor:
    print(x)

# Comprobamos tamaño
# check 

cursor.execute('select count(*) from puertos;')

for x in cursor:
    print(x)

#### Transit + Weather

df_transit_plus_weather = df_transit_plus_weather.rename(columns={'Unnamed: 0': 'transit_id'})
df_transit_plus_weather.head()

c('drop table if exists transit_plus_weather;')

query ='''

create table transit_plus_weather(
    transit_id int primary key not null,
    barco varchar(45),
    puerto_llegada varchar(45),
    LatitudPuertoLlegada float,
    LongitudPuertoLlegada float,
    llegada_fecha varchar(45),
    llegada_hora varchar(45),
    salida_fecha varchar(45),
    salida_hora varchar(45),
    daily_rain_sum float,
    daily_wind_speed_10m_max float
    );
'''

c(query)

# Relleno datos
nombre_tabla = 'transit_plus_weather'
nombre_columnas = ','.join(df_transit_plus_weather.columns) # creo una string

#bucle por filas
for i in range(df_transit_plus_weather.shape[0]):
    valores = tuple(df_transit_plus_weather.iloc[i].values)

    insert_query = f'insert into {nombre_tabla} ({nombre_columnas}) values {valores}'

    cursor.execute(insert_query)

insert_query

# Comprobamos

cursor.execute('select * from transit_plus_weather limit 2;')

for x in cursor:
    print(x)

# Comprobamos tamaño
# check 

cursor.execute('select count(*) from transit_plus_weather;')

for x in cursor:
    print(x)

#### OTLE GT 

df_otle_gt = df_otle_gt.rename(columns={'Unnamed: 0': 'otle_id'})
df_otle_gt.head()

# Voy a tener probelmas con los puntos de la columna 'gt' así que voy a quitar los puntos 
df_otle_gt['gt'] = df_otle_gt['gt'].str.replace('.', '')
df_otle_gt.head()

c('drop table if exists otle_gt;')

query ='''

create table otle_gt(
    otle_id int primary key not null,
    autoridad_portuaria	varchar(100),
    tipo_buque varchar(45),
    procedencia varchar(45),
    year int,
    num_buques int,
    gt float
    );
'''

c(query)

# Relleno datos
nombre_tabla = 'otle_gt'
nombre_columnas = ','.join(df_otle_gt.columns) # creo una string

#bucle por filas
for i in range(df_otle_gt.shape[0]):
    valores = tuple(df_otle_gt.iloc[i].values)

    insert_query = f'insert into {nombre_tabla} ({nombre_columnas}) values {valores}'

    cursor.execute(insert_query)

insert_query

# Comprobamos

cursor.execute('select * from otle_gt limit 2;')

for x in cursor:
    print(x)

# Comprobamos tamaño
# check 

cursor.execute('select count(*) from otle_gt;')

for x in cursor:
    print(x)

cursor.execute('show tables;')

for e in cursor:
    print(e)

* Me falta borrar una tabla que hice por error con una errata

table_name = 'trnasit_clean'
try:
    cursor.execute(f'DROP TABLE {table_name}')
    print('Table was dropped succesfully')
except:
    print('Error, table not found or already dropped beforehand.')


cursor.execute('show tables;')

for e in cursor:
    print(e)
